Win32 GUI Based RayTracing Feature 1. MemoryPool 2. ThreadPool 3. FieldObject   - Components(Transform, Resources, Material)

+++ Raytracing in a weekend/hitable_list.h
@@ -15,7 +15,7 @@ bool hitable_list::hit(const ray& r, float tmin, float tmax, hit_record& rec) co
 {
 	hit_record temp_rec;
 	bool hit_anything = false;
-	double closet_so_far = tmax;
+	float closet_so_far = tmax;
 
 	for (int i = 0; i < list_size; ++i)
 	{
